feat(dagwriter): seperate interface from implementation
move blockservice implementation to its own package
