+++ dagwriter.go
@@ -1,15 +1,9 @@
 package dagwriter
 
 import (
-	"bytes"
 	"context"
-	"fmt"
-	"io"
 
-	blocks "github.com/ipfs/go-block-format"
-	"github.com/ipfs/go-blockservice"
 	"github.com/ipld/go-ipld-prime"
-	cidlink "github.com/ipld/go-ipld-prime/linking/cid"
 )
 
 // DagWriter provides access for writing and deleting ipld.Nodes in an underlying store
@@ -36,55 +30,4 @@ type DagWritingService interface {
 	DagWriter
 	// NewBatchWriter instantiates a new multi-operation write/delete
 	NewBatchWriter() DagBatchWriter
-}
-
-type dagWritingService struct {
-	*ipld.LinkSystem
-	bs blockservice.BlockService
-}
-
-// NewDagWriter returns a new DagWritingService interface
-func NewDagWriter(bs blockservice.BlockService) DagWritingService {
-	ds := dagWritingService{bs: bs}
-	ls := cidlink.DefaultLinkSystem()
-	ls.StorageWriteOpener = ds.put
-	ds.LinkSystem = &ls
-	return ds
-}
-
-// ipld.BlockWriterOpener that writes to block store
-func (ds dagWritingService) put(lnkCtx ipld.LinkContext) (io.Writer, ipld.BlockWriteCommitter, error) {
-	buffer := new(bytes.Buffer)
-	committer := func(lnk ipld.Link) error {
-		asCidLink, ok := lnk.(cidlink.Link)
-		if !ok {
-			return fmt.Errorf("unsupported link type %v", lnk)
-		}
-		block, err := blocks.NewBlockWithCid(buffer.Bytes(), asCidLink.Cid)
-		if err != nil {
-			return err
-		}
-		return ds.bs.AddBlock(block)
-	}
-	return buffer, committer, nil
-}
-
-func (ds dagWritingService) Delete(ctx context.Context, lnk ipld.Link) error {
-	asCidLink, ok := lnk.(cidlink.Link)
-	if !ok {
-		return fmt.Errorf("unsupported link type %v", lnk)
-	}
-	return ds.bs.DeleteBlock(asCidLink.Cid)
-}
-
-func (ds dagWritingService) NewBatchWriter() DagBatchWriter {
-	dmw := &dagBatchWriter{bs: ds.bs, cache: newCachedOperationsStore()}
-	ls := ipld.LinkSystem{
-		StorageWriteOpener: dmw.put,
-		EncoderChooser:     ds.EncoderChooser,
-		DecoderChooser:     ds.DecoderChooser,
-		HasherChooser:      ds.HasherChooser,
-	}
-	dmw.LinkSystem = &ls
-	return dmw
 }
